pyproject.toml
[tool.poetry]
name = "electrohub"
version = "0.1.0"
description = "online platform-trading network of electronics"
authors = ["vitek <rgavih@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "^4.2"
djangorestframework = "^3.15.2"
black = "^24.10.0"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
